postgres/app/requirements.txt
pandas==2.1.4
psycopg2-binary==2.9.9
python-dotenv==1.0.1
streamlit==1.41.1
